Add new Application List Loader Loading all Application icons into the list might work on most devices, but in cases where we have 500/700+ apps installed, the device will crash with an outOfMemory exception.
To fix this we need to load as few icons as possible. We do this by adding a scroll loading feature that will make sure only to load and keep icons for currently visible views. At the same time we add a LruCache to store as many bitmaps as we can (without running out of memory) in order to reduce scroll lagging.