Only allow re-mapping on device attached keys We do not want to intercept external keyboards, mouses, gamepads etc.
Also try to load the default CharacterMap for the device attached keys. This might fix some issues on some device. At the same time we allow the injected flag to stay on while returning to the original methods. This should ensure that they don't start interfering with things like haptic feedback when we inject keys.