Disable original haptic feedback 'fix' This is a fix for the commit 902c0a2b
The first commit worked fine for long press actions, but since it was checking for the keydown flag and clicks are invoked on key up, it did not alays do what it should. This fix will provide a field which we can always use to check if one of the original methods is currently in the process of handling an event, during which time we will disable haptic.