Fix combo keyboard issue
When you press A and then B while keeping A pressed, Android will send an automatic up event on A before sending the down event on B. Because this happens at the same time, A does not have enough time to get from Queueing to Dispatching before B registers a new event and cancel A.

This is of cause an issue with the way the module handles keys while using a global shared object to keep track of them, and it has to be addressed. But as a temp work-around, we set a 1ms delay on combo down events, which will give primary events enough time to switch methods and start getting processed. The synchronized blocks will then take care of the rest.