Service: Bypass new SELinux restrictions The most important fix here is the service name change (Thanks to rovo89 for pointing it out).
Google has added name check to it's service control that is prefix based (like "user.*"). This is checked by SELinux when sending the binder to the native service manager. Service names without a valid prefix will not be excepted and the Service is not added. This might become even more strict with a more specific list of names in the future, but for the moment it will do the job.